requirements.txt
torch
transformers
peft
datasets
openai
google-generativeai
wandb
requests
google-api-python-client
IMDbPY
tqdm
python-dotenv
langdetect
scipy
vllm
google-genai
xai-sdk
openai-harmony
anthropic
pandas